requirements.txt
python-dotenv
altair==4.2.2
numpy==1.26.4
pandas==2.2.3
streamlit==1.39.0
streamlit-chat==0.1.1
mysqlclient==2.2.4
SQLAlchemy==2.0.35
mysql-connector-python==9.0.0
langchain==0.3.25
langchain-community==0.3.23
langchain_openai==0.3.16
langchain_chroma==0.2.3
langchain-ollama==0.3.2
pypdf==5.0.1
cryptography==43.0.1
matplotlib==3.9.2
streamlit-aggrid==1.0.5
japanize_matplotlib==1.1.3
wandb
datasets==3.0.1
evaluate==0.4.3
scikit-learn==1.5.2
packaging==24.2
Requests==2.32.3
wheel
transformers==4.51.2
#mpi4py==4.0.1
peft
sacrebleu==2.4.2
accelerate
bitsandbytes
nest_asyncio
beautifulsoup4
sentence-transformers
unstructured
openpyxl
readchar
openai
Levenshtein
ninja==1.11.1.4
langgraph
ipython
pillow
docling==2.20.0
docling-core==2.22.0
torchinfo==1.8.0